tests for rankedAdminnable and Whitelisting
